sh: Do not assume that SIGPIPE will only kill a subshell in builtins/wait3.0 test.
POSIX says that SIGPIPE affects a process and therefore a SIGPIPE caused and
received by a subshell environment may or may not affect the parent shell
environment.

The change assumes that ${SH} is executed in a new process. This must be the
case if it contains a slash and everyone appears to do so anyway even though
POSIX might permit otherwise.

This change makes builtins/wait3.0 work in ksh93.


git-svn-id: svn://svn.freebsd.org/base/head@236771 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f