Make the SoC stuff a little more modular, and start to move away from having the CPU device that's a child of atmelarm that does stuff.
o Create a linker_set for the support fucntions for the SoCs.
o Rename soc_data to soc_info.
o Move the delay and reset function pointers to new soc_data struct
o Create elements for all known SoCs
o Add lookup of the SoC we found, and print a warning if it isn't one
  we know about.


git-svn-id: svn://svn.freebsd.org/base/head@238376 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f