Rework time handling.
After I made the previous commit, I noticed the code does some things it
shouldn't. It casts a struct timeval to a time_t, assuming tv_sec is the
first member. Also, we are not interested in microseconds, so it is
better to just call time(NULL).

MFC after:	1 month


git-svn-id: svn://svn.freebsd.org/base/head@239988 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f